refactor(be): move s3 module to shared libs for client submission
