+++ apps/backend/apps/admin/src/admin.module.ts
@@ -18,6 +18,7 @@ import { AdminExceptionFilter, apolloErrorFormatter } from '@libs/exception'
 import { openTelemetryModuleOption } from '@libs/instrumentation'
 import { LoggingPlugin, pinoLoggerModuleOption } from '@libs/logger'
 import { PrismaModule } from '@libs/prisma'
+import { StorageModule } from '@libs/storage'
 import { NoticeModule } from '@admin/notice/notice.module'
 import { AdminController } from './admin.controller'
 import { AdminService } from './admin.service'
@@ -26,7 +27,6 @@ import { AssignmentModule } from './assignment/assignment.module'
 import { ContestModule } from './contest/contest.module'
 import { GroupModule } from './group/group.module'
 import { ProblemModule } from './problem/problem.module'
-import { StorageModule } from './storage/storage.module'
 import { SubmissionModule } from './submission/submission.module'
 import { UserModule } from './user/user.module'
 import { WorkbookModule } from './workbook/workbook.module'
