+++ apps/backend/apps/admin/src/problem/services/file.service.ts
@@ -10,7 +10,7 @@ import {
   UnprocessableFileDataException
 } from '@libs/exception'
 import { PrismaService } from '@libs/prisma'
-import { StorageService } from '@admin/storage/storage.service'
+import { StorageService } from '@libs/storage'
 import type { UploadFileInput } from '../model/problem.input'
 
 @Injectable()
