+++ apps/backend/libs/storage/src/content.type.ts
