+++ apps/backend/libs/storage/src/index.ts
@@ -1,0 +1,4 @@
+export * from './content.type'
+export * from './s3.provider'
+export * from './storage.module'
+export * from './storage.service'
