+++ apps/backend/libs/storage/src/s3.provider.ts
@@ -7,18 +7,36 @@ export const S3Provider = {
   provide: 'S3_CLIENT',
   import: [ConfigModule],
   inject: [ConfigService],
-  useFactory: async (config: ConfigService) =>
-    new S3Client({
+  useFactory: async (config: ConfigService) => {
+    // If production, assume that the service is running on AWS
+    // and use the default credentials
+    const isProduction = config.get('APP_ENV') === 'production'
+    if (isProduction) {
+      return new S3Client({
+        region: 'ap-northeast-2'
+      })
+    }
+
+    // TODO: integrate multiple AWS IAM access keys to single IAM user
+    // and do not separate TESTCASE and MEDIA
+    const accessKey = config.get<string>('TESTCASE_ACCESS_KEY')
+    const secretKey = config.get<string>('TESTCASE_SECRET_KEY')
+    if (!accessKey || !secretKey) {
+      throw new Error(
+        'TESTCASE_ACCESS_KEY or TESTCASE_SECRET_KEY is not defined'
+      )
+    }
+
+    return new S3Client({
       region: 'ap-northeast-2',
-      // TODO: production 환경에서는 endpoint, forcePathStyle 삭제
-      // endpoint: config.get('TESTCASE_ENDPOINT_URL'),
       endpoint: config.get('STORAGE_BUCKET_ENDPOINT_URL'),
       forcePathStyle: true,
       credentials: {
-        accessKeyId: config.get('TESTCASE_ACCESS_KEY') ?? '',
-        secretAccessKey: config.get('TESTCASE_SECRET_KEY') ?? ''
+        accessKeyId: accessKey,
+        secretAccessKey: secretKey
       }
     })
+  }
 }
 
 export const S3MediaProvider = {
